Use functions to construct regexes.
I'm not sure if this is the best direction, but I find that I can reason
easier about the regexes if I construct them using relatively simple code, rather than
typing everything by hand.

One thing that this fixes is that previously regexes would match
nonsensical, inconsistent dates such as "11-22/33". I think it makes
sense to require at least internally-consistent dates. If it's dashes,
it's got to be dashes, and not a mix.

I'm thinking that in the future the analyzer could also require
per-language consistency. But, one problem at the time. Today we're
focusing on issue#7.

Fixes #7.