Add a date index on top with links.
This way you can jump to a specific date.

I tried to make the index compact, by grouping dates per month. It can
be made prettier later, for now it's just the bare funcionality.

It looks like I attempted to do this some time in 2019, and didn't
manage to. Today I sat down and did it in maybe half an hour. It's
incredible how my cognitive abilities were hurt in 2019, and how long it
took to recover.

Resolves #6.